IBM-Swift/Kitura#888 Improve user profile for Facebook token
