+++ Sources/CredentialsFacebook/CredentialsFacebook.swift
@@ -28,7 +28,7 @@ import Foundation
 /// Authentication using Facebook web login with OAuth.
 /// See [Facebook's manual](https://developers.facebook.com/docs/facebook-login/manually-build-a-login-flow)
 /// for more information.
-public class CredentialsFacebook : CredentialsPluginProtocol {
+public class CredentialsFacebook: CredentialsPluginProtocol {
     
     private var clientId: String
     
@@ -67,7 +67,7 @@ public class CredentialsFacebook : CredentialsPluginProtocol {
     /// - Parameter clientSecret: The App Secret of the app in the Facebook Developer dashboard.
     /// - Parameter callbackUrl: The URL that Facebook redirects back to.
     /// - Parameter options: A dictionary of plugin specific options.
-    public init (clientId: String, clientSecret: String, callbackUrl: String, options: [String:Any]?=nil) {
+    public init(clientId: String, clientSecret: String, callbackUrl: String, options: [String:Any]?=nil) {
         self.clientId = clientId
         self.clientSecret = clientSecret
         self.callbackUrl = callbackUrl
@@ -89,11 +89,11 @@ public class CredentialsFacebook : CredentialsPluginProtocol {
     ///                     authentication data in the request.
     /// - Parameter inProgress: The closure to invoke to cause a redirect to the login page in the
     ///                     case of redirecting authentication.
-    public func authenticate (request: RouterRequest, response: RouterResponse,
-                              options: [String:Any], onSuccess: @escaping (UserProfile) -> Void,
-                              onFailure: @escaping (HTTPStatusCode?, [String:String]?) -> Void,
-                              onPass: @escaping (HTTPStatusCode?, [String:String]?) -> Void,
-                              inProgress: @escaping () -> Void) {
+    public func authenticate(request: RouterRequest, response: RouterResponse,
+                             options: [String:Any], onSuccess: @escaping (UserProfile) -> Void,
+                             onFailure: @escaping (HTTPStatusCode?, [String:String]?) -> Void,
+                             onPass: @escaping (HTTPStatusCode?, [String:String]?) -> Void,
+                             inProgress: @escaping () -> Void) {
         if let code = request.queryParameters["code"] {
             var requestOptions: [ClientRequest.Options] = []
             requestOptions.append(.schema("https://"))
@@ -131,7 +131,7 @@ public class CredentialsFacebook : CredentialsPluginProtocol {
                                         try profileResponse.readAllData(into: &body)
                                         jsonBody = JSON(data: body)
                                         if let dictionary = jsonBody.dictionaryObject,
-                                            let userProfile = self.createUserProfile(from: dictionary) {
+                                            let userProfile = createUserProfile(from: dictionary, for: self.name) {
                                             if let delegate = self.delegate {
                                                 delegate.update(userProfile: userProfile, from: dictionary)
                                             }
@@ -175,32 +175,4 @@ public class CredentialsFacebook : CredentialsPluginProtocol {
             }
         }
     }
-    
-    private func createUserProfile(from facebookData: [String:Any]) -> UserProfile? {
-        if let id = facebookData["id"] as? String,
-            let name = facebookData["name"] as? String {
-            
-            var userEmails: [UserProfile.UserProfileEmail]? = nil
-            if let email = facebookData["email"] as? String {
-                let userEmail = UserProfile.UserProfileEmail(value: email, type: "")
-                userEmails = [userEmail]
-            }
-            
-            var userName: UserProfile.UserProfileName? = nil
-            if let familyName = facebookData["familyName"] as? String,
-                let givenName = facebookData["givenName"] as? String {
-                let middleName = (facebookData["middleName"] as? String) ?? ""
-                userName = UserProfile.UserProfileName(familyName: familyName, givenName: givenName, middleName: middleName)
-            }
-            
-            var userPhotos: [UserProfile.UserProfilePhoto]? = nil
-            if let photo = facebookData["picture"] as? String {
-                let userPhoto = UserProfile.UserProfilePhoto(photo)
-                userPhotos = [userPhoto]
-            }
-            return UserProfile(id: id, displayName: name, provider: self.name, name: userName, emails: userEmails, photos: userPhotos)
-        }
-        return nil
-    }
-
 }
